+++ src/components/layout/SignedInLinks.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom'
 import { connect } from 'react-redux'
-import { signOut } from '../../store/actions/authAction'
+import { signOut } from '../../store/actions/authActions'
 
 const SignedInLinks = (props) => {
   return (
