--- src/store/actions/authAction.js
@@ -1,25 +1,0 @@
-export const signIn = (credentials) => {
-  return (dispatch, getState, {getFirebase}) => {
-    const firebase = getFirebase();
-
-    firebase.auth().signInWithEmailAndPassword(
-      credentials.email,
-      credentials.password
-    ).then(() => {
-      dispatch({type: 'LOGIN_SUCCESS'})
-    }).catch((err) => {
-      dispatch({ type: 'LOGIN_ERROR', err })
-    })
-  }
-}
-
-export const signOut = () => {
-  return (dispatch, getState, {getFirebase}) => {
-    const firebase = getFirebase();
-
-    firebase.auth().signOut().then(() => {
-      dispatch({ type: 'SIGNOUT_SUCCESS'});
-    });
-
-  }
-}
